refactor(admin): consolidate edit page logic and improve error handling
Merge the edit page logic into a single file and enhance error messages for better user feedback. Remove redundant code and streamline the UI components for consistency.
